BREAKING CHANGE: Adds user as foreign key to the DBF model
This is used to record the user that uploaded the file.

The break in the change is: I changed the initial migration to pretend the
model was always like this. This is not a problem since this was never a stable
branch and no database ever created this table without this foreignkey
constraint. I decided to do that because I didn't want to allow NULL values,
and I didn't want to create a default value for this field (since none would
make sense). If you did fork from this branch, make shure you make these
changes in your own database.