Adds simple DBF validation (with tests)
For now, the validation consists on trying to load the DBF file with dbfread.
If it fails to read the headers, we assume the file is not a valid DBF.

This includes changes to tests, since the validation now works for the dbf.
Because of that this commit includes test data that can be used (a valid and an
invalid dbf for now). This commit also makes all tests more alike when creating
DBF objects.