Fixes bug when validating DBF
We were reading the file from it's filename, but when we received an upload,
the file hadn't been commited to disk when `is_valid_dbf()` was called, so this
resulted in an error. We are now creating a tempfile for the content, since
dbfread only reads from a filename (it does not accept a file-like object), and
then deleting the tempfile.