Changes the container to use a real database instead of sqlite
For some reason we were using a sqlite database that was being reset
between rebuilds. This has worked until now only because we didn't
actually use django models (and the managed database) at all. Now that
we needed it, I ran into this. I decided to keep the database separate
from the one we use for the data just to keep things simpler (though you
can actually use the same one; just point to it).