WIP: adds chunked_upload views and endpoints
This is still broken in, at least, the following ways:
    - I removed the original DBF upload form
    - The typeahead plugin is broken (so the search does not work).
    - The upload serves no purpose at all for now.

The reason I commited this is to have a history of what it took to make
the chunked_upload widget work by itself, before integrating with the
dbf upload functionality itself. Otherwise I'd have a huge commit, which
would be really hard for someone just looking at it to understand.