WIP: Puts together the chunked_upload widget and the DBF upload form
This uses the chunked upload file and the metadata the user inputs in
the form to create a DBF object and save it to our database.

It is still a work in progress because we still need to polish the
widget UI and to fix some issues (such as the typeahead and datepicker).