Warns user that the form has unsave data that will be lost if they leave the page
We need this because the form UI was not clear enough so the user would
know that the progress bar reaching 100% is not enough to have the data
imported. We improved the message in 8988055 but this adds another layer
of protection to make sure the user does not loose data.