Adds example of a broken SINAN file with wrong data types
We got some SINAN files with wrong data types in what should be DATE
fields. This file has the wrong structure we got in a real example, but
with all the fields with the correct name

WARNING: tests are broken, in part because of the chunked upload
changes, in part because of the removal of geodjango (which left
migrations that require postgis to be the database engine). We need to
fix them before merging this code.