Removes expensive validation from the document upload process
The idea is that we will need this validation when importing the file
anyways. To avoid taking too long to process the request for the file
upload, we'll only keep this initial validation to the headers of the
DBF file. At a later time, when the import process runs (assyncronously
to this request/response cycle), we will inform the user about status
of the import.