Installs geopandas via apt on dockers
The version of geopandas that's being installed via pip ends up trying
to install subprocess32, which should only be installed on python2. This
does not happen with the debian package.
Beyond that, geopandas is a big dependency and has a package available
in debian, so it sounds like a good idea to install it through apt (the
same way we install pandas itself).